requirement.txt
twitchio==2.1.0
pyautogui==0.9.53
python-dotenv==1.0.0